feat(locking): Add AuthorityFromUID
The support of custom authorities is already there in methods
`AuthenticateLockingSP` and `AuthenticateAdminSP`, e.g.:
https://github.com/open-source-firmware/go-tcg-storage/blob/87417253ae2dfec902dd4995ef36c2d9a48d41d1/pkg/locking/locking.go#L87

But there is no way provided to build an `*authority` with
a non-empty value of `auth` field.

In this commit I add a function that constructs an `*authority`
with `auth` field filled with an uid.AuthorityObjectUID value.

This is useful if package `pkg/locking` is used by other software
that has theirown ideas which authority should be used when.

Signed-off-by: Dmitrii Okunev <xaionaro@meta.com>